improve played track handling
A problem occurred when we starting playing alternatives. We were
adding the alternative track id to the played tracks, but were only
checking the original track id to see if it had been played before.

We now do all the played track handling in the Spotify client so that
we can do all the adding and checking on the resolved track.